AI_Interview/requirements.txt
streamlit
mistralai
sentence-transformers
faiss-cpu
pypdf
numpy